move code compile function into separate src solidity support, state tests new field "solidity" new code keywords ":raw 0x11,  :solidity contractName"
